requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
SQLAlchemy==1.4.7
Werkzeug==1.0.1
python-dotenv
Flask-Migrate